# [quantization] NHWC specialization for quantized::cat