--- a/aten/src/ATen/cpu/vec256/vec256_qint.h
+++ b/aten/src/ATen/cpu/vec256/vec256_qint.h
@@ -206,7 +206,7 @@ struct Vec256<c10::qint8> {
   float_vec_return_type dequantize(
       Vec256<float> scale,
       Vec256<float> zero_point,
-      Vec256<float> scale_zp_premul) const {
+      Vec256<float> scale_neg_zp_premul) const {
     __m128i int_val0 = _mm_set1_epi64x(_mm256_extract_epi64(vals, 0));
     __m128i int_val1 = _mm_set1_epi64x(_mm256_extract_epi64(vals, 1));
     __m128i int_val2 = _mm_set1_epi64x(_mm256_extract_epi64(vals, 2));
@@ -219,13 +219,13 @@ struct Vec256<c10::qint8> {
 
 #if defined(__AVX2__) && defined(__FMA__)
     auto val0 =
-        vec256::fmadd(scale, Vec256<float>(float_val0), scale_zp_premul);
+        vec256::fmadd(scale, Vec256<float>(float_val0), scale_neg_zp_premul);
     auto val1 =
-        vec256::fmadd(scale, Vec256<float>(float_val1), scale_zp_premul);
+        vec256::fmadd(scale, Vec256<float>(float_val1), scale_neg_zp_premul);
     auto val2 =
-        vec256::fmadd(scale, Vec256<float>(float_val2), scale_zp_premul);
+        vec256::fmadd(scale, Vec256<float>(float_val2), scale_neg_zp_premul);
     auto val3 =
-        vec256::fmadd(scale, Vec256<float>(float_val3), scale_zp_premul);
+        vec256::fmadd(scale, Vec256<float>(float_val3), scale_neg_zp_premul);
 #else
     auto val0 = scale * (Vec256<float>(float_val0) - zero_point);
     auto val1 = scale * (Vec256<float>(float_val1) - zero_point);