--- a/test/test_quantized.py
+++ b/test/test_quantized.py
@@ -598,9 +598,8 @@ class TestQuantizedOps(TestCase):
                                               min_side=1, max_side=10),
                        qparams=hu.qparams()),
            relu=st.booleans(),
-           out_scale=st.floats(-0.5, 0.5),
-           out_zero_pt=st.integers(1, 126))
-    def test_cat_nhwc(self, X, relu, out_scale, out_zero_pt):
+           out_qparams=hu.qparams())
+    def test_cat_nhwc(self, X, relu, out_qparams):
         X, (scale, zero_point, torch_type) = X
         # Tile out X so # channels is > 64
         X = np.repeat(X, 70 / X.shape[3], 3)
@@ -609,13 +608,16 @@ class TestQuantizedOps(TestCase):
         qX = torch.quantize_linear(X, scale, zero_point, torch_type).permute([0, 3, 1, 2])
         qY = torch.quantize_linear(Y, scale, zero_point, torch_type).permute([0, 3, 1, 2])
 
+        out_scale, out_zero_point, _ = out_qparams
+
         ref = torch.cat([qX.dequantize(), qY.dequantize()], dim=1)
         if relu:
             ref[ref < 0] = 0.0
         ref = torch.quantize_linear(ref, scale=out_scale, zero_point=out_zero_pt, dtype=torch_type)
 
         if relu:
-            out = torch.ops.quantized.cat_relu([qX, qY], dim=1, scale=out_scale, zero_point=out_zero_pt)
+            out = torch.ops.quantized.cat_relu(
+                [qX, qY], dim=1, scale=out_scale, zero_point=out_zero_pt)
         else:
             out = torch.ops.quantized.cat([qX, qY], dim=1, scale=out_scale, zero_point=out_zero_pt)