--- a/test/test_quantized.py
+++ b/test/test_quantized.py
@@ -594,6 +594,7 @@ class TestQuantizedOps(TestCase):
                              zero_point=zero_point)
 
     """Tests quantize concatenation (both fused and not)."""
+    @no_deadline
     @given(X=hu.tensor(shapes=hu.array_shapes(min_dims=4, max_dims=4,
                                               min_side=1, max_side=10),
                        qparams=hu.qparams()),
@@ -601,15 +602,15 @@ class TestQuantizedOps(TestCase):
            out_qparams=hu.qparams())
     def test_cat_nhwc(self, X, relu, out_qparams):
         X, (scale, zero_point, torch_type) = X
+        out_scale, out_zero_pt, out_torch_type = out_qparams
+        assume(out_torch_type == torch_type)
+
         # Tile out X so # channels is > 64
         X = np.repeat(X, 70 / X.shape[3], 3)
         X = torch.from_numpy(np.ascontiguousarray(X))
         Y = torch.neg(X)
         qX = torch.quantize_linear(X, scale, zero_point, torch_type).permute([0, 3, 1, 2])
         qY = torch.quantize_linear(Y, scale, zero_point, torch_type).permute([0, 3, 1, 2])
-
-        out_scale, out_zero_pt, out_torch_type = out_qparams
-        assume(out_torch_type == torch_type)
 
         ref = torch.cat([qX.dequantize(), qY.dequantize()], dim=1)
         if relu: