--- a/test/test_quantized.py
+++ b/test/test_quantized.py
@@ -598,30 +598,28 @@ class TestQuantizedOps(TestCase):
     @given(X=hu.tensor(shapes=hu.array_shapes(min_dims=4, max_dims=4,
                                               min_side=1, max_side=10),
                        qparams=hu.qparams()),
-           relu=st.booleans(),
-           out_qparams=hu.qparams())
-    def test_cat_nhwc(self, X, relu, out_qparams):
+           relu=st.booleans())
+    def test_cat_nhwc(self, X, relu):
         X, (scale, zero_point, torch_type) = X
-        out_scale, out_zero_pt, out_torch_type = out_qparams
-        assume(out_torch_type == torch_type)
 
         # Tile out X so # channels is > 64
         X = np.repeat(X, 70 / X.shape[3], 3)
-        X = torch.from_numpy(np.ascontiguousarray(X))
-        Y = torch.neg(X)
+        X = torch.from_numpy(np.ascontiguousarray(X)).float()
+        Y = X.clone()
+        Y = torch.from_numpy(np.ascontiguousarray(Y))
         qX = torch.quantize_linear(X, scale, zero_point, torch_type).permute([0, 3, 1, 2])
         qY = torch.quantize_linear(Y, scale, zero_point, torch_type).permute([0, 3, 1, 2])
 
         ref = torch.cat([qX.dequantize(), qY.dequantize()], dim=1)
         if relu:
             ref[ref < 0] = 0.0
-        ref = torch.quantize_linear(ref, scale=out_scale, zero_point=out_zero_pt, dtype=torch_type)
+        ref = torch.quantize_linear(ref, scale=scale, zero_point=zero_point, dtype=torch_type)
 
         if relu:
             out = torch.ops.quantized.cat_relu(
-                [qX, qY], dim=1, scale=out_scale, zero_point=out_zero_pt)
+                [qX, qY], dim=1, scale=scale, zero_point=zero_point)
         else:
-            out = torch.ops.quantized.cat([qX, qY], dim=1, scale=out_scale, zero_point=out_zero_pt)
+            out = torch.ops.quantized.cat([qX, qY], dim=1, scale=scale, zero_point=zero_point)
 
         torch.testing.assert_allclose(out.dequantize(), ref.dequantize())
         self.assertNotEqual(out.stride(), sorted(out.stride()))